feature/multi-device-sync: refactor parsers
